+++ addons/alexander/justech_alexander_reports/hooks.py
@@ -18,3 +18,17 @@ def post_init_hook(env):
         if vals:
             company.write(vals)
     env["ir.actions.report"]._dx_attach_invoice_edi_pdf()
+    leftover = (
+        env["ir.actions.report"]
+        .sudo()
+        .search(
+            [
+                (
+                    "report_name",
+                    "=",
+                    "justech_alexander_reports.report_saleorder_conduce",
+                )
+            ]
+        )
+    )
+    leftover.unlink()